added cli outside docker

+++ api_controller.py
@@ -38,11 +38,10 @@ def as_form(cls: Type[BaseModel]):
 
 @as_form
 class Params(BaseModel):
-    quality_factor: int
+    quality_factor: float
     file_mode: str
     check_intervals: int
     keywords: list[str]
-
 
 
 app = FastAPI()
@@ -86,9 +85,8 @@ async def receive(file: UploadFile = File(...), form: Params = Depends(Params.as
         f.close()
 
     ocr_service.edit_photo(list_of_concerns_double, 'photo_under_procces.png',
-                     quality_factor)
+                           quality_factor)
     return FileResponse('store_output_photo/photo_under_procces.png',  media_type="photo/png")
 
 if __name__ == '__main__':
     uvicorn.run(app, host='0.0.0.0', port=8005)
-
